+++ src/components/icon-lang/index.tsx
@@ -1,44 +1,41 @@
 import i18n from "i18next";
 import { Dropdown, Menu } from "antd";
-import { useState } from "react";
-import Button from "../../Button";
+import { BsTranslate } from "react-icons/bs";
+import { useTranslation } from "react-i18next";
 
-const Languages = () => {
-  const [language, setLanguage] = useState("vi");
-
+const IconLanguages = () => {
+  const { t } = useTranslation();
   const menu = (
     <Menu>
-      <Menu.Item>
+      <Menu.Item key='vi'>
         <p
+          className='lang__item'
           onClick={() => {
             i18n.changeLanguage("vi");
-            setLanguage("vi");
-            console.log("vi");
           }}
         >
-          Tiếng việt
+          {t("default.language.vi")}
         </p>
       </Menu.Item>
-      <Menu.Item>
+      <Menu.Item key='en'>
         <p
+          className='lang__item'
           onClick={() => {
             i18n.changeLanguage("en");
-            setLanguage("en");
-            console.log("en");
           }}
         >
-          Tiếng anh
+          {t("default.language.en")}
         </p>
       </Menu.Item>
     </Menu>
   );
   return (
-    <div className='cws-lang'>
+    <div className='lang'>
       <Dropdown overlay={menu} placement='bottomRight'>
-        <Button>{language}</Button>
+        <BsTranslate />
       </Dropdown>
     </div>
   );
 };
 
-export default Languages;
+export default IconLanguages;