+++ src/locales/en/page.ts
@@ -1,5 +1,0 @@
-export const TranslationEn = {
-  home: {
-    title: "Home",
-  },
-};