+++ src/locales/en/pages.ts
@@ -1,0 +1,14 @@
+export const TranslationEnPages = {
+  home: {
+    title: "Home",
+  },
+  login: {
+    title: "Login",
+    email: "Email",
+    password: "Password",
+    forgotPass: "Forgot password?",
+    notMember: "Not a member?",
+    registerNow: "Register now.",
+    or: "or",
+  },
+};