+++ src/locales/vi/page.ts
@@ -1,5 +1,0 @@
-export const TranslationVi = {
-  home: {
-    title: "Trang chủ",
-  },
-};