+++ src/locales/vi/pages.ts
@@ -1,0 +1,14 @@
+export const TranslationViPages = {
+  home: {
+    title: "Trang chủ",
+  },
+  login: {
+    title: "Đăng nhập",
+    email: "Email",
+    password: "Mật khẩu",
+    forgotPass: "Quên mật khẩu?",
+    notMember: "Chưa là thành viên?",
+    registerNow: "Đăng ký ngay.",
+    or: "hoặc",
+  },
+};