+++ src/pages/home/Home.tsx
@@ -1,9 +1,4 @@
-import { useTranslation } from "react-i18next";
-import LoginForm from "../../components/login/loginForm";
-
 const Home = () => {
-  const { t } = useTranslation();
-
   return <section></section>;
 };