+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Các thay đổi:
 
+## 10/03/2022
+
 - [ui] add login page
+- [feature] attach i18 in login page
+- [feature] validate input text
+- [feature] change language by i18
 
 ## 09/03/2022
 
 - [improve] Setup redux toolkit
+- Remove alias import
 
 ## 03/03/2022